refactor: move into screaming folders
