+++ src/common/entity/replace_reference_with_alias.py
