+++ src/tests/bdd/steps/domain.py
@@ -3,8 +3,8 @@ import json
 from behave import given, then
 
 from authentication.models import AccessControlList
+from common.entity.create_entity import CreateEntity
 from common.tree.tree_node import ListNode, Node
-from common.utils.create_entity import CreateEntity
 from common.utils.get_storage_recipe import storage_recipe_provider
 from domain_classes.blueprint_attribute import BlueprintAttribute
 from enums import SIMOS, BuiltinDataTypes
