+++ src/tests/unit/common/utils/test_create_default_array.py
@@ -1,7 +1,10 @@
 import unittest
 
-from common.utils.arrays import create_default_array, remove_first_and_join
-from common.utils.create_entity import CreateEntity
+from common.entity.create_entity import CreateEntity
+from common.entity.create_entity_arrays import (
+    create_default_array,
+    remove_first_and_join,
+)
 from domain_classes.blueprint import Blueprint
 from domain_classes.blueprint_attribute import BlueprintAttribute
 from domain_classes.dimension import Dimension
