+++ src/tests/unit/common/utils/test_create_entity.py
@@ -1,6 +1,6 @@
 import unittest
 
-from common.utils.create_entity import CreateEntity
+from common.entity.create_entity import CreateEntity
 from domain_classes.blueprint_attribute import BlueprintAttribute
 from tests.unit.mock_data.mock_blueprint_provider import MockBlueprintProvider
 from tests.unit.mock_data.mock_document_service import get_mock_document_service
